refactor(frontend): Better organize the location of component files for the data page
