+++ frontend/src/lib/components/data/compare_view/CompareFilters.svelte
@@ -4,17 +4,17 @@
     import * as AlertDialog from "$lib/components/ui/alert-dialog/index.js";
 	import { Buildings, Calendar, Faders, Info, PlusCircle, Selection, Users, X } from "phosphor-svelte";
 	import { onMount } from "svelte";
-	import FilterList from "./FilterList.svelte";
-	import Button from "../ui/button/button.svelte";
+	import FilterList from "../FilterList.svelte";
+	import Button from "../../ui/button/button.svelte";
 	import { getSeasonsSeasonsGet } from "$lib/api/seasons/seasons";
 	import type { GetDataFiltersDataFiltersGetParams, SeasonResponse } from "$lib/api/model";
 	import { getDataFiltersDataFiltersGet } from "$lib/api/data/data";
 	import { toast } from "svelte-sonner";
-	import Separator from "../ui/separator/separator.svelte";
-	import SelectMatchDialog from "./SelectMatchDialog.svelte";
-	import BaseDialog from "../generic/dialogs/BaseDialog.svelte";
-	import EventList from "../generic/event_list/EventList.svelte";
-    import type { Filters as EventListFilters } from "../generic/event_list/EventList.svelte";
+	import Separator from "../../ui/separator/separator.svelte";
+	import SelectMatchDialog from "./select_match_dialog/SelectMatchDialog.svelte";
+	import BaseDialog from "../../generic/dialogs/BaseDialog.svelte";
+	import EventList from "../../generic/event_list/EventList.svelte";
+    import type { Filters as EventListFilters } from "../../generic/event_list/EventList.svelte";
     import { db } from "$lib/utils/db";
 
     let { filters = $bindable(), fields } = $props();
