+++ frontend/src/lib/components/data/compare_view/charts/CompareCapability.svelte
