+++ frontend/src/lib/components/data/compare_view/charts/CompareScore.svelte
