+++ frontend/src/lib/components/data/data_view/TeamOthers.svelte
