+++ frontend/src/lib/components/data/data_view/displays/CapabilityDisplay.svelte
