+++ frontend/src/lib/components/data/data_view/displays/ScoreDisplay.svelte
